+++ src/features/ui/ToCsvButton.tsx
@@ -31,11 +31,13 @@ export function ToCsvButton({
   filename,
   label = 'Export All Data to CSV',
 }: ToCsvProps) {
+  console.info('data', data)
+
   if (!data || data.length === 0) return null
 
   return (
     <button
-      className="flex h-[40px] w-full flex-col items-center justify-center rounded border-2 border-secondary-500 bg-secondary-700 p-2 px-4 py-2 text-sm font-bold text-white text-white/90 hover:border-secondary-700 hover:bg-secondary-500 hover:text-white disabled:bg-gray-500 sm:h-[60px] md:w-[150px]"
+      className="flex h-[40px] w-full max-w-[200px] flex-col items-center justify-center rounded border-2 border-secondary-500 bg-secondary-700 p-2 px-4 py-2 text-sm font-bold text-white text-white/90 hover:border-secondary-700 hover:bg-secondary-500 hover:text-white disabled:bg-gray-500 sm:h-[60px] md:min-w-[150px]"
       aria-label="Export All Data to CSV"
       onClick={() => toCsv(data, filename)}
     >